refactor : cartUtil함수로 분리
